backend/requirements.txt
appdirs==1.4.4
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
fs==2.4.16
inflection==0.5.1
Markdown==3.10
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
unicode-slugify==0.1.5
Unidecode==1.4.0
